Moves most of docker content to root
This was done so that I can add pip install step to the image build.
If I left the content in the .devcontainer folder then I would have to
copy the setup.py into the .devcontainer. Making 2 copies of it. 2
places to change the deps.

the .. to . and slight changes to dir paths are key here